moved all utils from cli.py to cli_utils.py
so they can be used from plugins without causing circular imports.
